added FormFileHelper and modified CategoryController

+++ Controllers/CategoryController.cs
@@ -1,7 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using SistemasWeb01.Helpers;
 using SistemasWeb01.Models;
 using SistemasWeb01.Repository.Implementations;
+using SistemasWeb01.ViewModels;
 using static System.Collections.Specialized.BitVector32;
 
 namespace SistemasWeb01.Controllers
@@ -9,9 +11,11 @@ namespace SistemasWeb01.Controllers
     public class CategoryController : Controller
     {
         private readonly ICategoryRepository _categoryRepository;
-        public CategoryController(ICategoryRepository categoryRepository)
+        private readonly IFormFileHelper _formFileHelper;
+        public CategoryController(ICategoryRepository categoryRepository, IFormFileHelper formFileHelper)
         {
             _categoryRepository = categoryRepository;
+            _formFileHelper = formFileHelper;
         }
         public IActionResult Index()
         {
@@ -20,17 +24,22 @@ namespace SistemasWeb01.Controllers
         }
         public IActionResult Create()
         {
-            return View();
+            CategoryViewModel categoryViewModel = new CategoryViewModel();
+            return View(categoryViewModel);
         }
 
         [HttpPost]
-        public IActionResult Create(Category category)
+        public async Task<IActionResult> Create(CategoryViewModel categoryViewModel)
         {
             if (ModelState.IsValid)
             {
                 try
                 {
-                    _categoryRepository.CreateCategory(category);
+                    if (categoryViewModel.formFile != null)
+                    {
+                        categoryViewModel.ThumbnailImage = await _formFileHelper.UploadFile(categoryViewModel.formFile);
+                    }
+                    _categoryRepository.CreateCategory(categoryViewModel);
                     TempData["mensaje"] = "La categoría se creó correctamente";
                     return RedirectToAction("Index");
                 }
@@ -51,7 +60,7 @@ namespace SistemasWeb01.Controllers
                     ModelState.AddModelError(string.Empty, exception.Message);
                 }
             }
-            return View(category);
+            return View(categoryViewModel);
         }
 
         public IActionResult Details(int id)
@@ -62,6 +71,7 @@ namespace SistemasWeb01.Controllers
                 return NotFound();
             }
             return View(category);
+            
         }
 
         public IActionResult Edit(int id)
@@ -71,20 +81,36 @@ namespace SistemasWeb01.Controllers
             {
                 return NotFound();
             }
-            return View(category);
+            CategoryViewModel categoryViewModel = new()
+            {
+                Id = category.Id,
+                Name = category.Name,
+                ThumbnailImage = category.ThumbnailImage
+
+            };
+            return View(categoryViewModel);
         }
 
         [HttpPost]
-        public IActionResult Edit(Category category)
+        public async Task<IActionResult> Edit(CategoryViewModel categoryViewModel)
         {
             if (ModelState.IsValid)
             {
                 try
                 {
-                    _categoryRepository.EditCategory(category);
+                    if (categoryViewModel.formFile != null)
+                    {
+                        if (categoryViewModel.ThumbnailImage != null)
+                        {
+                            _formFileHelper.DeleteFile(categoryViewModel.ThumbnailImage);
+                        }
+                        categoryViewModel.ThumbnailImage = await _formFileHelper.UploadFile(categoryViewModel.formFile);
+                    }
+                    _categoryRepository.EditCategory(categoryViewModel);
                     TempData["mensaje"] = "La categoría se actualizó correctamente";
-                    return RedirectToAction(nameof(Index));
+                    return RedirectToAction("Index");
                 }
+
                 catch (DbUpdateException dbUpdateException)
                 {
                     if (dbUpdateException.InnerException.Message.Contains("UNIQUE constraint failed: Categories.Name"))
@@ -101,7 +127,7 @@ namespace SistemasWeb01.Controllers
                     ModelState.AddModelError(string.Empty, exception.Message);
                 }
             }
-            return View(category);
+            return View(categoryViewModel);
         }
 
         public IActionResult Delete(int id)
@@ -117,6 +143,9 @@ namespace SistemasWeb01.Controllers
         [HttpPost]
         public IActionResult Delete(Category category)
         {
+            if (category == null) return NotFound();
+            _formFileHelper.DeleteFile(category.ThumbnailImage);
+
             _categoryRepository.DeleteCategory(category);
             TempData["mensaje"] = "La categoría se eliminó correctamente";
             return RedirectToAction("Index");
